move align_to function to wgpu_types crate
