+++ wgpu-hal/src/dx12/command.rs
@@ -46,7 +46,7 @@ impl crate::BufferTextureCopy {
                             (self.size.width / block_width) * block_size
                         }
                     };
-                    crate::auxil::align_to(actual, d3d12_ty::D3D12_TEXTURE_DATA_PITCH_ALIGNMENT)
+                    wgt::math::align_to(actual, d3d12_ty::D3D12_TEXTURE_DATA_PITCH_ALIGNMENT)
                 },
             },
         }
